PR comments: move code, remove deprecated methods
